+++ packages/client/src/form/FormError.tsx
@@ -1,38 +1,43 @@
-import { translator } from "@use-pico/common";
+import { type ErrorSchema, translator } from "@use-pico/common";
 import type { FC } from "react";
 import { ErrorIcon } from "../icon/ErrorIcon";
 import { Icon } from "../icon/Icon";
 import { FormErrorCls } from "./FormErrorCls";
 
 export namespace FormError {
+	export interface Meta {
+		isDirty: boolean;
+		isTouched: boolean;
+		errors: ErrorSchema.Type[];
+	}
+
 	export interface Props extends FormErrorCls.Props {
-		errors: string[] | undefined;
+		meta: Meta;
 	}
 }
 
 export const FormError: FC<FormError.Props> = ({
-	errors,
+	meta,
 	variant,
 	tva = FormErrorCls,
 	cls,
 }) => {
 	const { slots } = tva(variant, cls);
 
-	if (!errors || errors.length === 0) {
-		return null;
-	}
+	const shouldShowError =
+		meta.isDirty && meta.isTouched && meta.errors && meta.errors.length > 0;
 
-	return (
+	return shouldShowError ? (
 		<div className={slots.base()}>
-			{errors.map((error, index) => (
+			{meta.errors.map((error, index) => (
 				<div
 					key={`${index}-${error}`}
 					className={slots.error()}
 				>
 					<Icon icon={ErrorIcon} />
-					<span>{translator.rich(error)}</span>
+					<span>{translator.rich(error.message)}</span>
 				</div>
 			))}
 		</div>
-	);
+	) : null;
 };